Use size_t instead of long long for gradient_structure::CMPDIF_BUFFER_SIZE to avoid conversion warnings.

+++ src/linad99/gradstrc.cpp
@@ -89,7 +89,7 @@ long int gradient_structure::GRADSTACK_BUFFER_SIZE = 4000000L;
 long int gradient_structure::CMPDIF_BUFFER_SIZE=140000000L;
 #else
 long long int gradient_structure::GRADSTACK_BUFFER_SIZE = 4000000L;
-long long int gradient_structure::CMPDIF_BUFFER_SIZE=140000000L;
+size_t gradient_structure::CMPDIF_BUFFER_SIZE=140000000L;
 #endif
 
 dependent_variables_information * gradient_structure::DEPVARS_INFO=NULL;
